Move messages/src/error to messages/src/utils/error
