+++ messages/src/connection/invite.rs
@@ -1,4 +1,4 @@
-use crate::error::prelude::*;
+use crate::utils::error::prelude::*;
 use crate::out_of_band::invitation::OutOfBandInvitation;
 use crate::a2a::{A2AMessage, MessageId};
 use crate::connection::did::Did;
