+++ messages/src/issuance/credential.rs
@@ -1,4 +1,4 @@
-use crate::error::MessagesResult;
+use crate::utils::error::MessagesResult;
 use crate::a2a::{A2AMessage, MessageId};
 use crate::ack::please_ack::PleaseAck;
 use crate::attachment::{AttachmentId, Attachments};
