+++ messages/src/lib.rs
@@ -23,7 +23,6 @@ pub mod attachment;
 pub mod basic_message;
 pub mod connection;
 pub mod discovery;
-pub mod error;
 pub mod problem_report;
 pub mod forward;
 pub mod issuance;
