+++ messages/src/out_of_band/invitation.rs
@@ -1,5 +1,5 @@
 use crate::a2a_message;
-use crate::error::prelude::*;
+use crate::utils::error::prelude::*;
 use crate::a2a::message_type::MessageType;
 use crate::a2a::{A2AMessage, MessageId};
 use crate::attachment::Attachments;
