+++ messages/src/proof_presentation/presentation_proposal.rs
@@ -1,4 +1,4 @@
-use crate::error::prelude::*;
+use crate::utils::error::prelude::*;
 use crate::a2a::message_family::MessageFamilies;
 use crate::a2a::message_type::MessageType;
 use crate::a2a::{A2AMessage, MessageId};
