+++ messages/src/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod validation;
 pub mod devsetup;
+pub mod error;
